iter18: add tests
