+++ internal/handler/shorten_test.go
@@ -1,0 +1,88 @@
+package handler_test
+
+import (
+	"io"
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"testing"
+
+	"github.com/gin-gonic/gin"
+	"github.com/google/uuid"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestShortenURL(t *testing.T) {
+
+	type request struct {
+		httpMethod  string
+		requestBody io.Reader
+	}
+
+	tests := []struct {
+		testName                string
+		request                 request
+		expectedRespCode        int
+		expectedRespContentType string
+	}{
+		{
+			testName: "test wrong http method",
+			request: request{
+				httpMethod:  http.MethodGet,
+				requestBody: nil,
+			},
+			expectedRespCode: http.StatusBadRequest,
+		},
+		{
+			testName: "test empty body",
+			request: request{
+				httpMethod:  http.MethodPost,
+				requestBody: nil,
+			},
+			expectedRespCode: http.StatusBadRequest,
+		},
+		{
+			testName: "test correct request",
+			request: request{
+				httpMethod:  http.MethodPost,
+				requestBody: strings.NewReader(testURL + "test"),
+			},
+			expectedRespCode: http.StatusCreated,
+		},
+	}
+
+	gin.SetMode(gin.TestMode)
+
+	for _, test := range tests {
+		t.Run(test.testName, func(t *testing.T) {
+			w := httptest.NewRecorder()
+			c, _ := gin.CreateTestContext(w)
+
+			c.Request = httptest.NewRequest(test.request.httpMethod, "/", test.request.requestBody)
+
+			t.Logf("sending %s request to %s", c.Request.Method, c.Request.RequestURI)
+			userID := uuid.NewString()
+			c.Set("user_id", userID)
+
+			hs.ShortenURL(c)
+
+			result := w.Result()
+
+			body := result.Body
+			defer body.Close() //nolint:errcheck
+
+			bodyData, err := io.ReadAll(body)
+			statusCode := result.StatusCode
+
+			require.NoError(t, err)
+
+			if statusCode == http.StatusCreated {
+				t.Logf("response body is: %v", string(bodyData))
+				assert.NotEmpty(t, bodyData)
+			}
+
+			assert.Equal(t, test.expectedRespCode, statusCode)
+		})
+	}
+}
