+++ pkg/gzip/gzip.go
@@ -1,4 +1,4 @@
-package middleware
+package gzip
 
 import (
 	"io"
@@ -9,21 +9,21 @@ import (
 // gzipWriter is a custom wrapper around gin.ResponseWriter that enables gzip compression
 // for HTTP responses. It embeds the original ResponseWriter, adds an io.Writer for
 // compressed output, and tracks the HTTP status code.
-type gzipWriter struct {
+type GzipWriter struct {
 	gin.ResponseWriter
-	writer     io.Writer
+	Writer     io.Writer
 	statusCode int
 }
 
 // Write writes the provided byte slice to the underlying gzip writer.
 // It returns the number of bytes written and any error encountered during the write operation.
-func (gzw *gzipWriter) Write(p []byte) (int, error) {
-	return gzw.writer.Write(p)
+func (gzw *GzipWriter) Write(p []byte) (int, error) {
+	return gzw.Writer.Write(p)
 }
 
 // WriteHeader sets the HTTP status code for the response and writes it to the underlying ResponseWriter.
 // It also stores the status code in the gzipWriter for later reference.
-func (gzw *gzipWriter) WriteHeader(statusCode int) {
+func (gzw *GzipWriter) WriteHeader(statusCode int) {
 	gzw.statusCode = statusCode
 	gzw.ResponseWriter.WriteHeader(gzw.statusCode)
 }
